add services

+++ src/main/kotlin/com/team13/junction/model/SensorData.kt
@@ -1,12 +1,13 @@
 package com.team13.junction.model
 
 import com.fasterxml.jackson.annotation.JsonAlias
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class SensorData(
-    @JsonAlias("Consumption") val consumption: Double,
-    @JsonAlias("Temp") val temp: Double,
-    @JsonAlias("FlowTime") val flowTime: Double,
-    @JsonAlias("Power_Consumption") val powerConsumption: Double,
+    @JsonAlias("Consumption") val consumption: BigDecimal,
+    @JsonAlias("Temp") val temp: BigDecimal,
+    @JsonAlias("FlowTime") val flowTime: BigDecimal,
+    @JsonAlias("Power_Consumption") val powerConsumption: BigDecimal,
     @JsonAlias("TimeStamp") val timeStamp: LocalDateTime,
 )
